rework arithmetic example
