+++ examples/Arithmetic/demo_dynamic.cpp
