+++ examples/Arithmetic/demo_simple.cpp
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "basic_nodes.hpp"
+#include "nodes.hpp"
 #include <iostream>
 
 using namespace geoflow;
